Add driver code block and rework all_paths
Added a code block to test the all_paths function and printing
readability. Reworked the all_paths function to simply just output
all routes that lead to the desired ending position instead of a
root node of the tree. Changed the length of the history array needed
until a branch is cut short.